fix: prevent money dupe

+++ src/utils/functions/economy/offers.ts
@@ -78,6 +78,9 @@ export async function setBlockedList(userId: string, list: string[]) {
 }
 
 export async function deleteOffer(offer: Offer, client: NypsiClient) {
+  offer = await prisma.offer.findUnique({ where: { messageId: offer.messageId } });
+  if (offer.sold) return false;
+
   await prisma.offer.delete({ where: { messageId: offer.messageId } });
 
   await updateBalance(offer.ownerId, (await getBalance(offer.ownerId)) + Number(offer.money));
